Add files via upload
Write a program to read an employee's data (name,
gross salary and tax). Then show the data from the
employee (name and net salary). Then raise the salary
of the employee based on a given percentage (only the
gross salary is affected by the percentage) and re-display the
employee data. Use the designed class below.